```
from asv_bench.benchmarks.io.sas import *
self8 = SAS()
self8.setup("sas7bdat")

%timeit self8.time_read_sas("sas7bdat")
41.6 ms ± 1.77 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
22.7 ms ± 1.04 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR

```